wip secretary reorg
